+++ src/components/HarvestingConfiguration/HarvestingConfigurationView.js
@@ -16,7 +16,6 @@ import {
   Row,
 } from '@folio/stripes/components';
 
-import { COUNTER } from '../../util/constants';
 import extractHarvesterImpls, { isServiceTypeSupported } from '../../util/harvesterImpls';
 import { AggregatorInfoView } from './AggregatorInfo';
 import { SushiCredentialsView } from './SushiCredentials';
@@ -65,9 +64,6 @@ const HarvestingConfigurationView = ({
     requestedReports = reports.join(', ');
   }
 
-  const counterVersion = get(usageDataProvider, 'harvestingConfig.reportRelease', '');
-  const reportReleaseLabel = counterVersion ? `${COUNTER} ${counterVersion}` : <NoValue />;
-
   const harvestingStart = usageDataProvider.harvestingConfig?.harvestingStart ?? <NoValue />;
   const harvestingEnd = usageDataProvider.harvestingConfig?.harvestingEnd ?? <NoValue />;
 
@@ -78,7 +74,7 @@ const HarvestingConfigurationView = ({
         <Col xs={3}>
           <KeyValue
             label={<FormattedMessage id="ui-erm-usage.udpHarvestingConfig.reportRelease" />}
-            value={reportReleaseLabel}
+            value={usageDataProvider.harvestingConfig?.reportRelease}
           />
         </Col>
         <Col xs={3}>